test: verify every native import's prototype against the headers (B2)
The prototype probe was hand-written and checked four routes out of 861. It is
now generated from CNA.Interop.Native by reflection, so PROTO_IMPORTS equals
PROTO_VERIFIED equals 861 and PROTO_UNMAPPABLE is zero.

Each import becomes a file-scope function pointer declared with the prototype
derived from the managed declaration and initialised with the real C function.
C's compatibility rules then turn a wrong return type, arity, parameter type,
by-ref direction or pointer depth into a diagnostic, and -Werror turns a
diagnostic into a failed gate. The C compiler is the authority on what the
header means; nothing here reimplements C's type rules.

It found four real defects on its first run, all fixed here.
cna_game_components_insert was declared with a signed 32-bit index against a
uint64_t parameter -- only x86-64's zero-extending 32-bit moves made that work,
and a negative index would have become an enormous unsigned one. The three
draw_*_primitives routes passed int for a CNA_PrimitiveType, which is uint32_t.
A fifth followed: CnaUserPrimitives.PrimitiveType was int against a uint32_t
field, which B1's offset probe cannot see because the widths agree. Struct field
signedness is therefore still unmeasured, and that is recorded as the next step.

109 parameters are listed in an explicit override manifest in four categories,
none of which is an ABI difference: const (C# cannot qualify a pointer, and
const is not part of any calling convention), char* (C's char is a third type
distinct from signed and unsigned char, so no C# pointer is exact), void*
against byte*, and 24 callbacks the managed side declares nint, which carries no
signature at all. Every entry was produced by the compiler rather than by hand:
the generator emitted the managed-derived prototype, the compiler rejected it,
and the recorded type is the one the diagnostic named. The manifest therefore
cannot excuse a difference that was never measured, and any other change to the
same parameter still fails to compile.

Nine negative controls run as part of the gate and all nine are rejected: wrong
return type, signedness at the same width, pointer depth, in->out, out->in,
wrong versioned descriptor, wrong callback shape, swapped same-width parameters,
absent import. A control whose target declaration is no longer generated reports
itself stale instead of passing silently -- which is what all nine did when they
were first written from memory rather than from the generated text. The gate was
also proven able to fail: making one mutation identical to its original yields
ABI_STATUS=failed.

Two tool defects fixed on the way. Run() read one pipe to the end and then the
other, which deadlocks as soon as a child fills the pipe nobody is draining; a C
compiler with a hundred diagnostics does exactly that, and the first full run
hung silently until killed. And the self-test ran after the report was built, so
its findings could not reach ABI_STATUS.

Verify-Abi.sh now defaults to ../../cnanext rather than the older ../../cna
checkout, which is a different ABI generation and must not be authoritative by
adjacency. Stale "854 imports" counts corrected to 861 across plan/README/docs.

api-compat 257/257, 0 diagnostics. 225 + 560 + 140 tests. ABI policy passed.

+++ src/CNA.Framework/GameComponentCollection.cs
@@ -101,7 +101,11 @@ public class GameComponentCollection : ICollection<GameComponent>
 
         _known[item.NativeHandle.Value] = item;
 
-        CnaResult result = Native.cna_game_components_insert(GameHandle, index, item.NativeHandle);
+        // The C index is a uint64_t. This passed a signed 32-bit value until B2's generated
+        // prototype probe compared the two: the argument register's upper half was never written,
+        // which happens to work on x86-64 because a 32-bit move zero-extends, and is not something
+        // to rely on. The negative case is refused above, so the widening is total.
+        CnaResult result = Native.cna_game_components_insert(GameHandle, (ulong)index, item.NativeHandle);
         if (result.IsFailure())
         {
             _known.Remove(item.NativeHandle.Value);
